feat(Wallet) complete the filter API
It uses the current data only and doesn't extend with new types or
include new features in activity sources DBs.

Major changes:
- Partially filter by chain IDs
- Partially filter by Status if it is the case
- Partially filter by token types
- Filter by counterparty addresses
- Use wallet accounts for TO/FROM instead of filters

Closes: #10634

+++ services/wallet/transfer/database_test.go
@@ -223,8 +223,8 @@ func TestGetTransfersForIdentities(t *testing.T) {
 	require.Equal(t, big.NewInt(trs[3].BlkNumber), entries[1].BlockNumber)
 	require.Equal(t, uint64(trs[1].Timestamp), entries[0].Timestamp)
 	require.Equal(t, uint64(trs[3].Timestamp), entries[1].Timestamp)
-	require.Equal(t, trs[1].ChainID, entries[0].NetworkID)
-	require.Equal(t, trs[3].ChainID, entries[1].NetworkID)
+	require.Equal(t, uint64(trs[1].ChainID), entries[0].NetworkID)
+	require.Equal(t, uint64(trs[3].ChainID), entries[1].NetworkID)
 	require.Equal(t, MultiTransactionIDType(0), entries[0].MultiTransactionID)
 	require.Equal(t, MultiTransactionIDType(0), entries[1].MultiTransactionID)
 }
